Extract out start out further
